fixed distance lore, used CustomConfig from infinitylib
